Update tests based on Layout.build
